requirements.txt
yfinance
python-dotenv
